move benchmark functionality to swiftlint/Helpers/Benchmark.swift
