+++ Source/swiftlint/Commands/LintCommand.swift
@@ -12,34 +12,6 @@ import Result
 import SourceKittenFramework
 import SwiftLintFramework
 
-private let numberFormatter: NSNumberFormatter = {
-    let formatter = NSNumberFormatter()
-    formatter.numberStyle = .DecimalStyle
-    formatter.minimumFractionDigits = 3
-    return formatter
-}()
-
-private let timestamp: String = {
-    let formatter = NSDateFormatter()
-    formatter.dateFormat = "yyyy_MM_dd_HH_mm_ss"
-    return formatter.stringFromDate(NSDate())
-}()
-
-private func saveBenchmark(name: String, times: [(id: String, time: Double)]) {
-    let string = times
-        .reduce([String: Double](), combine: { accu, idAndTime in
-            var accu = accu
-            accu[idAndTime.id] = (accu[idAndTime.id] ?? 0) + idAndTime.time
-            return accu
-        })
-        .sort({ $0.1 < $1.1 })
-        .map({ "\(numberFormatter.stringFromNumber($0.1)!): \($0.0)" })
-        .joinWithSeparator("\n")
-        + "\n"
-    let data = string.dataUsingEncoding(NSUTF8StringEncoding)
-    data?.writeToFile("benchmark_\(name)_\(timestamp).txt", atomically: true)
-}
-
 struct LintCommand: CommandType {
     let verb = "lint"
     let function = "Print lint warnings and errors (default command)"
